perf(seo): create SEO component
This component was created to improve SEO performance, by adding extra meta tags on header. It was
separeted on a component called SEO. In this component was added Open Graph/Facebook and twitter
meta tags. We still got to add a better image, because it has an Alura one. We also created the
robots.txt document on Public , to direct Google's SEO better.