refactor(provider): refactor pages to create a single page wrapper and providers
In this commit, we made a basic separation of Next and React scopes. First, we separated de Home
Screen from the Home Page. To make this avaiable, we had to bring the toggleModalCAdastro function
to the React context structure, so Modal's state could be read on Menu and Homepage. Then we created
the WebsitePageWrapper component, that contains the modal context provider. This component also
contains the common layout of the page: header, modal and footer. To remove the necessity of
repeating the content of the WebsitePageWrapper, we created the websitePageHOC, that receives a
component and its props, inside a WebsiteGlobalProvider (containing theme and style providers) and
the WebsitePageWrapper itself. At last, we refactorated all pages adjusting to the new structure.