feat(faq): add content to FAQ page from Alura's API
We created a "dynamic-static" page using Next's structure to populate FAQ answers page. Using the
API's page slug, we obtained the question's answer and category's other questions. We also changed
the websiteHOC component to receive pageWrapperProps from the component. This allowed the answer to
provide the Head Title of the page.