fix(filename): change the name of a file so Git can correct its path
I had changed the name of the folder from "Form" to "form", but Git didn't understood as a change.
But when I pulled the commit, it acused that filepath was incorrect, because of the case sensitive
path recognition. If I changed the folder name to uppercase, Git wouldn't track the change, so I
changed it to singular. Later I might change again to the correct name.